Game Breaking Interact FIx
- added an enum to interactable called INTERACTABLETYPE
which holds the different types of interactables.
Basically acts as a tag

- interact now holds a reference to the kid instead of
 just using the currentInteractable

- basically I put in a bunch of if statements in interact
to make sure if you're holding a child you can't
 interact with anything else, and you can only interact
with the child if that reference to it isn't null.

- even if the child has ran away from you while you're
still trying to capture it the game doesn't break.
I need to make it so once it gets away from
you the minigame stops

- Fixed light detect bug where even if the light wasn't on
it would still calculate percent visibility